Refactor: extract components from `/projects/new` page
- The page became bloated
