+++ frontend/src/components/NewProject/Sync/index.module.scss
@@ -1,0 +1,11 @@
+.urlInput {
+  min-width: 300px;
+  max-height: 37px;
+  flex-grow: 1;
+  margin-bottom: 20px;
+  max-width: 400px;
+}
+
+.syncButton {
+  margin-left: 20px !important;
+}
