+++ frontend/src/pages/projects/new/index.module.scss
